+++ nilearn/input_data/nifti_labels_masker.py
@@ -214,6 +214,10 @@ class NiftiLabelsMasker(BaseMasker, CacheMixin):
 
             mask_data, mask_affine = masking._load_mask_img(self.mask_img_)
 
+        if not hasattr(self, '_resampled_labels_img_'):
+            # obviates need to run .transform() before .inverse_transform()
+            self._resampled_labels_img_ = self.labels_img_
+
         return self
 
     def fit_transform(self, imgs, confounds=None):